[refactor] 챔피언 검색창 컴포넌트로 구조 변경
